knvram: Kernel-based NVRAM device driver
This adds knvram, a kernel based NVRAM device driver.

With knvram, the kernel implement a typical NVRAM device, with a
run-time in-memory shadow image of the backend NVRAM image. It is
synchronized to backend image on request. For boards with a small UPS,
and NMI on power-failure, it can be integrated as a reliable NVRAM
device.

Backend support is currently limited to memory mapped, NVRAM like
devices, and is tested with FRAM. Plan is to add support for at least
UBI. MTD backend could also be added, but UBI adds wear-leveling, so
is strongly preferred.

The backend image can be partitioned, and each partition can
optionally be exported to user-space as a character device.

Finally, there is transaction support. So that multiple write's can be
merged to an atomic commit.

Signed-off-by: Esben Haabendal <eha@deif.com>