appwd: wdt_gpio: Add pulse-width-ns support
Add support for specifying pulse-width-ns in devicetree.  When set, it
defines the pulse with (in nanoseconds) that should be generated on each
heartbeat.  The polarity of the pulse (active low/high) of the pulse is
defined by the gpios binding flags.  When pulse-width-ns is not set, the
level is toggled on each heartbeat.

Signed-off-by: Esben Haabendal <esben.haabendal@prevas.dk>