appwd: imx2_wdt: add restart_handler
We should register a restart_handler so that the board gets properly
rebooted/powered off. This is mostly cosmetic, as the watchdog will
tear us down anyway, but it's nice to avoid

  Reboot failed -- System halted

The code is more or less just stolen from the actual imx2 watchdog
driver and backported to fit into our copy.